feat: drop support for transpiling the code that is being prevaled
It cause way too many weird bugs that take a TON of time to track down.

BREAKING CHANGE: Code that's being prevaled will no longer be transpiled. You should either transpile it ahead of time or only write the code in a version of JavaScript that runs natively in the version of node you're running.